Refactor ResponsiveCanvas component
This commit refactors the `ResponsiveCanvas` component to align with the project's unified responsive architecture principles as outlined in the component refactor plan.

The key changes include:
- Removing all JavaScript-based device detection for UI rendering and replacing it with a CSS-first approach.
- Decomposing the monolithic `ResponsiveCanvas` component into smaller, more focused components (`CanvasContainer`, `SlideRenderer`, `ElementRenderer`).
- Extracting complex logic into custom hooks (`useCanvasDimensions`, `useCanvasGestures`, `useElementInteractions`).
- Creating a new directory structure under `src/client/components/slides/` with `canvas/` and `hooks/` subdirectories to better organize the new files.

This refactoring improves maintainability, separates concerns, and adheres to the strict architectural rules of the project. All existing tests pass after these changes.

Also updated `COMPONENT_REFACTOR_PLAN.md` to mark Phase 2.1 as complete.
