+++ COMPONENT_REFACTOR_PLAN.md
@@ -37,7 +37,13 @@ if (deviceType === 'mobile') {
 
 #### **2.1 ResponsiveCanvas.tsx Complete Refactor**
 
-**Current Violations to Fix**:
+**Current State**: ✅ Completed
+- Refactored to remove all JavaScript-based device detection for UI rendering.
+- Decomposed the monolithic component into smaller, more focused components (`CanvasContainer`, `SlideRenderer`, `ElementRenderer`).
+- Extracted complex logic into custom hooks (`useCanvasDimensions`, `useCanvasGestures`, `useElementInteractions`).
+- Moved all related files to a new `src/client/components/slides/canvas/` directory structure.
+
+**Original Violations Fixed**:
 1. **Device Detection for UI Logic** (Lines 199, 217, 220, etc.)
    ```typescript
    // REMOVE THIS PATTERN:
